LayoutTests/imported/w3c/web-platform-tests/largest-contentful-paint/idlharness-expected.txt



FAIL idl_test setup promise_test: Unhandled rejection with value: "Timed out waiting for LargestContentfulPaint entry"
PASS idl_test validation
PASS LargestContentfulPaint includes PaintTimingMixin: member names are unique
PASS Element includes ParentNode: member names are unique
PASS Element includes NonDocumentTypeChildNode: member names are unique
PASS Element includes ChildNode: member names are unique
PASS Element includes Slottable: member names are unique
PASS LargestContentfulPaint interface: existence and properties of interface object
PASS LargestContentfulPaint interface object length
PASS LargestContentfulPaint interface object name
PASS LargestContentfulPaint interface: existence and properties of interface prototype object
PASS LargestContentfulPaint interface: existence and properties of interface prototype object's "constructor" property
PASS LargestContentfulPaint interface: existence and properties of interface prototype object's @@unscopables property
PASS LargestContentfulPaint interface: attribute loadTime
PASS LargestContentfulPaint interface: attribute size
PASS LargestContentfulPaint interface: attribute id
PASS LargestContentfulPaint interface: attribute url
PASS LargestContentfulPaint interface: attribute element
PASS LargestContentfulPaint interface: operation toJSON()
FAIL LargestContentfulPaint must be primary interface of lcp assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: lcp"
FAIL Stringification of lcp assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: lcp"
FAIL LargestContentfulPaint interface: lcp must inherit property "loadTime" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: lcp"
FAIL LargestContentfulPaint interface: lcp must inherit property "size" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: lcp"
FAIL LargestContentfulPaint interface: lcp must inherit property "id" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: lcp"
FAIL LargestContentfulPaint interface: lcp must inherit property "url" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: lcp"
FAIL LargestContentfulPaint interface: lcp must inherit property "element" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: lcp"
FAIL LargestContentfulPaint interface: lcp must inherit property "toJSON()" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: lcp"
FAIL LargestContentfulPaint interface: default toJSON operation on lcp undefined is not an object (evaluating 'memberHolderObject.toJSON')
